+++ lib/component/template/battle/battle.dart
